Move fileUpload logic to attachment model
